Refactor SDK queries and mutations handling
