+++ src/@sdk/api/Cart/index.ts
@@ -1,7 +1,7 @@
 import { ErrorListener } from "@sdk/helpers";
 import { JobsManager } from "@sdk/jobs";
 import { ErrorCartTypes } from "@sdk/jobs/Cart";
-import { CheckoutNetworkManager } from "@sdk/network";
+import { NetworkManager } from "@sdk/network";
 import { CheckoutRepositoryManager, ICheckoutModel } from "@sdk/repository";
 import { SaleorState } from "@sdk/state";
 import { ISaleorStateSummeryPrices, StateItems } from "@sdk/state/types";
@@ -28,12 +28,12 @@ export class SaleorCartAPI extends ErrorListener implements ISaleorCartAPI {
 
   private checkoutRepositoryManager: CheckoutRepositoryManager;
   private saleorState: SaleorState;
-  private checkoutNetworkManager: CheckoutNetworkManager;
+  private networkManager: NetworkManager;
   private jobsManager: JobsManager;
 
   constructor(
     checkoutRepositoryManager: CheckoutRepositoryManager,
-    checkoutNetworkManager: CheckoutNetworkManager,
+    networkManager: NetworkManager,
     saleorState: SaleorState,
     loadOnStart: boolean,
     jobsManager: JobsManager
@@ -41,7 +41,7 @@ export class SaleorCartAPI extends ErrorListener implements ISaleorCartAPI {
     super();
     this.saleorState = saleorState;
     this.checkoutRepositoryManager = checkoutRepositoryManager;
-    this.checkoutNetworkManager = checkoutNetworkManager;
+    this.networkManager = networkManager;
     this.jobsManager = jobsManager;
 
     this.loaded = false;
@@ -110,7 +110,7 @@ export class SaleorCartAPI extends ErrorListener implements ISaleorCartAPI {
       const {
         data,
         error,
-      } = await this.checkoutNetworkManager.getRefreshedCheckoutLines(
+      } = await this.networkManager.getRefreshedCheckoutLines(
         this.saleorState.checkout.lines
       );
 
@@ -144,7 +144,7 @@ export class SaleorCartAPI extends ErrorListener implements ISaleorCartAPI {
       const {
         data,
         error,
-      } = await this.checkoutNetworkManager.getRefreshedCheckoutLines(
+      } = await this.networkManager.getRefreshedCheckoutLines(
         this.saleorState.checkout.lines
       );
 
@@ -179,7 +179,7 @@ export class SaleorCartAPI extends ErrorListener implements ISaleorCartAPI {
       const {
         data,
         error,
-      } = await this.checkoutNetworkManager.getRefreshedCheckoutLines(
+      } = await this.networkManager.getRefreshedCheckoutLines(
         this.saleorState.checkout.lines
       );
 
@@ -214,7 +214,7 @@ export class SaleorCartAPI extends ErrorListener implements ISaleorCartAPI {
       const {
         data,
         error,
-      } = await this.checkoutNetworkManager.getRefreshedCheckoutLines(
+      } = await this.networkManager.getRefreshedCheckoutLines(
         this.saleorState.checkout.lines
       );
