+++ src/@sdk/react/queries.ts
@@ -11,13 +11,15 @@ export const useUserDetails = queryFactory("getUserDetails");
  * @deprecated Use useCheckout hook instead
  */
 export const useUserCheckout = queryFactory("getUserCheckout");
-
 /**
  * @deprecated Use useCheckout hook instead
  */
 export const useCheckoutDetails = queryWithVariablesFactory(
   "getCheckoutDetails"
 );
+/**
+ * @deprecated Use useCheckout hook instead
+ */
 export const useCheckoutProductVariants = queryWithVariablesFactory(
   "getCheckoutProductVariants"
 );
